syncing with efforts for AccuracyCoin compat

+++ src/opcodes/unofficial/axa.h
@@ -4,7 +4,7 @@
 
 //------------------------------------------------------------------------------
 // Name: opcode_axa
-// Desc: AND X register with accumulator then AND result with 7 and store in memory.
+// Desc: X & A & M.hi
 // Alias: SHA, AHX
 //------------------------------------------------------------------------------
 struct opcode_axa {
@@ -15,7 +15,7 @@ struct opcode_axa {
 		// NOTE(eteran): we don't need an explicit +1, because
 		// it is automatically caused by the data fetch
 		const uint8_t value = (address >> 8) & X & A;
-		address             = (address & 0x00ff) | (value << 8);
+		address             = (address & 0x00ff) | (static_cast<uint16_t>(value) << 8);
 		return value;
 	}
 };
